Project done

+++ database/migrations/2020_07_21_184255_create_libris_table.php
@@ -15,7 +15,7 @@ class CreateLibrisTable extends Migration
     {
         Schema::create('libris', function (Blueprint $table) {
             $table->id();
-            $table->integer('ISBN');
+            $table->bigInteger('ISBN');
             $table->string('titulli');
             $table->integer('stoku');
             $table->string('foto');
